requirements.txt
# tensorflow=1.14
tensorflow-gpu=1.14